requirements.txt
python-telegram-bot==13.7
redis==3.5.3